c backend: improve float handling

+++ tests/machine/x/c/python_math.c
@@ -8,22 +8,22 @@ extern double math_pow(double x, double y);
 extern double math_sin(double x);
 extern double math_log(double x);
 
-static double r = 3;
+static double r = 3.0;
 
 int main() {
-  int area = 3.141592653589793 * (r * r);
+  double area = 3.141592653589793 * (r * r);
   double root = 7;
   double sin45 = __builtin_sin(3.141592653589793 / 4.0);
   double log_e = __builtin_log(2.718281828459045);
   printf("%s ", "Circle area with r =");
-  printf("%d ", r);
+  printf("%.17g ", r);
   printf("%s ", "=>");
-  printf("%d\n", area);
+  printf("%.17g\n", area);
   printf("%s ", "Square root of 49:");
-  printf("%.16g\n", root);
+  printf("%.17g\n", root);
   printf("%s ", "sin(π/4):");
-  printf("%.16g\n", sin45);
+  printf("%.17g\n", sin45);
   printf("%s ", "log(e):");
-  printf("%.16g\n", log_e);
+  printf("%.17g\n", log_e);
   return 0;
 }
